pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "arxivscribe"
version = "1.0.0"
description = "AI-powered arXiv paper digests — fetch, summarize, search, and organize research papers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Abel Yagubyan", email = "abelo9996@gmail.com"}]
keywords = ["arxiv", "papers", "research", "ai", "summarization", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.4.0"]
all = ["ollama>=0.4.0"]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.21.0", "pytest-cov>=4.1.0"]

[project.urls]
Homepage = "https://github.com/Abelo9996/ArxivScribe"
Repository = "https://github.com/Abelo9996/ArxivScribe"
Issues = "https://github.com/Abelo9996/ArxivScribe/issues"

[project.scripts]
arxivscribe = "arxivscribe.cli:cli"

[tool.setuptools.packages.find]
include = ["arxivscribe*"]

[tool.setuptools.package-data]
"arxivscribe" = ["web/templates/**/*", "web/static/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
